Move probs functos to MeasuresFunctors.hpp
