+++ pennylane_lightning/core/src/simulators/lightning_kokkos/measurements/ExpValFunctors.hpp
@@ -15,8 +15,8 @@
 #include <Kokkos_Core.hpp>
 
 #include "BitUtil.hpp"
-#include "BitUtilKokkos.hpp"
 #include "Error.hpp"
+#include "UtilKokkos.hpp"
 
 /// @cond DEV
 namespace {
@@ -270,321 +270,6 @@ template <class PrecisionT> struct getExpectationValueSparseFunctor {
     void operator()(const std::size_t row, PrecisionT &expval) const {
         for (std::size_t j = indptr[row]; j < indptr[row + 1]; j++) {
             expval += real(conj(arr[row]) * data[j] * arr[indices[j]]);
-        }
-    }
-};
-
-template <class PrecisionT, class DeviceType> class getProbsFunctor {
-  public:
-    // Required for functor:
-    using execution_space = DeviceType;
-    using value_type = PrecisionT[];
-    const unsigned value_count;
-
-    using ComplexT = Kokkos::complex<PrecisionT>;
-    Kokkos::View<ComplexT *> arr;
-    Kokkos::View<std::size_t *> all_indices;
-    Kokkos::View<std::size_t *> all_offsets;
-    getProbsFunctor(const Kokkos::View<ComplexT *> &arr_,
-                    [[maybe_unused]] const std::vector<std::size_t> &wires_,
-                    const Kokkos::View<std::size_t *> all_indices_,
-                    const Kokkos::View<std::size_t *> all_offsets_)
-        : value_count{1U << wires_.size()}, arr{arr_},
-          all_indices{all_indices_}, all_offsets{all_offsets_} {}
-
-    KOKKOS_INLINE_FUNCTION
-    void init(PrecisionT dst[]) const {
-        for (unsigned i = 0; i < value_count; ++i)
-            dst[i] = 0.0;
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void join(PrecisionT dst[], const PrecisionT src[]) const {
-        for (unsigned i = 0; i < value_count; ++i)
-            dst[i] += src[i];
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void operator()(const std::size_t i, const std::size_t j,
-                    PrecisionT dst[]) const {
-        const std::size_t index = all_indices(i) + all_offsets(j);
-        const PrecisionT rsv = arr(index).real();
-        const PrecisionT isv = arr(index).imag();
-        dst[i] += rsv * rsv + isv * isv;
-    }
-};
-
-template <class PrecisionT, class DeviceType, std::size_t num_wires>
-class getProbsNQubitOpFunctor {
-  public:
-    // Required for functor:
-    using execution_space = DeviceType;
-    using value_type = PrecisionT[];
-    const unsigned value_count;
-
-    using UnmanagedSizeTHostView =
-        Kokkos::View<std::size_t *, Kokkos::HostSpace,
-                     Kokkos::MemoryTraits<Kokkos::Unmanaged>>;
-    using ComplexT = Kokkos::complex<PrecisionT>;
-    using KokkosComplexVector = Kokkos::View<ComplexT *>;
-    Kokkos::View<ComplexT *> arr;
-    const std::size_t n_wires;
-
-    Kokkos::View<std::size_t *> parity;
-    Kokkos::View<std::size_t *> rev_wires;
-
-    std::size_t rev_wire_0;
-    std::size_t rev_wire_1;
-    std::size_t rev_wire_2;
-    std::size_t rev_wire_3;
-    std::size_t rev_wire_4;
-    std::size_t rev_wire_5;
-    std::size_t rev_wire_6;
-    std::size_t rev_wire_7;
-    std::size_t parity_0;
-    std::size_t parity_1;
-    std::size_t parity_2;
-    std::size_t parity_3;
-    std::size_t parity_4;
-    std::size_t parity_5;
-    std::size_t parity_6;
-    std::size_t parity_7;
-    std::size_t parity_8;
-
-    getProbsNQubitOpFunctor(
-        const Kokkos::View<ComplexT *> &arr_, const std::size_t num_qubits_,
-        [[maybe_unused]] const std::vector<std::size_t> &wires_)
-        : value_count{1U << wires_.size()}, arr{arr_}, n_wires{wires_.size()} {
-        PL_ABORT_IF(num_wires != 0 && num_wires != n_wires,
-                    "num_wires must be equal to n_wires.");
-        const std::size_t n_wires = wires_.size();
-        std::vector<std::size_t> rev_wires_(n_wires);
-        for (std::size_t k = 0; k < n_wires; k++) {
-            rev_wires_[n_wires - 1 - k] = (num_qubits_ - 1) - wires_[k];
-        }
-        std::vector<std::size_t> parity_ =
-            Pennylane::Util::revWireParity(rev_wires_);
-        if constexpr (num_wires == 0) {
-            Kokkos::resize(rev_wires, rev_wires_.size());
-            Kokkos::deep_copy(
-                rev_wires,
-                UnmanagedSizeTHostView(rev_wires_.data(), rev_wires_.size()));
-            Kokkos::resize(parity, parity_.size());
-            Kokkos::deep_copy(
-                parity, UnmanagedSizeTHostView(parity_.data(), parity_.size()));
-        }
-        if constexpr (num_wires > 0) {
-            rev_wire_0 = rev_wires_[0];
-            parity_0 = parity_[0];
-            parity_1 = parity_[1];
-        }
-        if constexpr (num_wires > 1) {
-            rev_wire_1 = rev_wires_[1];
-            parity_2 = parity_[2];
-        }
-        if constexpr (num_wires > 2) {
-            rev_wire_2 = rev_wires_[2];
-            parity_3 = parity_[3];
-        }
-        if constexpr (num_wires > 3) {
-            rev_wire_3 = rev_wires_[3];
-            parity_4 = parity_[4];
-        }
-        if constexpr (num_wires > 4) {
-            rev_wire_4 = rev_wires_[4];
-            parity_5 = parity_[5];
-        }
-        if constexpr (num_wires > 5) {
-            rev_wire_5 = rev_wires_[5];
-            parity_6 = parity_[6];
-        }
-        if constexpr (num_wires > 6) {
-            rev_wire_6 = rev_wires_[6];
-            parity_7 = parity_[7];
-        }
-        if constexpr (num_wires > 7) {
-            rev_wire_7 = rev_wires_[7];
-            parity_8 = parity_[8];
-        }
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void init(PrecisionT dst[]) const {
-        for (unsigned i = 0; i < value_count; ++i)
-            dst[i] = 0.0;
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void join(PrecisionT dst[], const PrecisionT src[]) const {
-        for (unsigned i = 0; i < value_count; ++i)
-            dst[i] += src[i];
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply0(const std::size_t i0, const std::size_t rev_wire,
-                const std::size_t offset, PrecisionT dst[]) const {
-        std::size_t i1;
-        i1 = i0 | (0U << rev_wire);
-        PrecisionT rsv = real(arr(i1));
-        PrecisionT isv = imag(arr(i1));
-        dst[offset + 0] += rsv * rsv + isv * isv;
-        i1 = i0 | (1U << rev_wire);
-        rsv = real(arr(i1));
-        isv = imag(arr(i1));
-        dst[offset + 1] += rsv * rsv + isv * isv;
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply1(const std::size_t i0, const std::size_t rev_wire_0,
-                const std::size_t rev_wire_1, const std::size_t offset,
-                PrecisionT dst[]) const {
-        apply0(i0, rev_wire_0, 0 + offset, dst);
-        apply0(i0 | (1U << rev_wire_1), rev_wire_0, 2 + offset, dst);
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply2(const std::size_t i0, const std::size_t rev_wire_0,
-                const std::size_t rev_wire_1, const std::size_t rev_wire_2,
-                const std::size_t offset, PrecisionT dst[]) const {
-        apply1(i0, rev_wire_0, rev_wire_1, 0 + offset, dst);
-        apply1(i0 | (1U << rev_wire_2), rev_wire_0, rev_wire_1, 4 + offset,
-               dst);
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply3(const std::size_t i0, const std::size_t rev_wire_0,
-                const std::size_t rev_wire_1, const std::size_t rev_wire_2,
-                const std::size_t rev_wire_3, const std::size_t offset,
-                PrecisionT dst[]) const {
-        apply2(i0, rev_wire_0, rev_wire_1, rev_wire_2, 0 + offset, dst);
-        apply2(i0 | (1U << rev_wire_3), rev_wire_0, rev_wire_1, rev_wire_2,
-               8 + offset, dst);
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply4(const std::size_t i0, const std::size_t rev_wire_0,
-                const std::size_t rev_wire_1, const std::size_t rev_wire_2,
-                const std::size_t rev_wire_3, const std::size_t rev_wire_4,
-                const std::size_t offset, PrecisionT dst[]) const {
-        apply3(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3, 0 + offset,
-               dst);
-        apply3(i0 | (1U << rev_wire_4), rev_wire_0, rev_wire_1, rev_wire_2,
-               rev_wire_3, 16 + offset, dst);
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply5(const std::size_t i0, const std::size_t rev_wire_0,
-                const std::size_t rev_wire_1, const std::size_t rev_wire_2,
-                const std::size_t rev_wire_3, const std::size_t rev_wire_4,
-                const std::size_t rev_wire_5, const std::size_t offset,
-                PrecisionT dst[]) const {
-        apply4(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3, rev_wire_4,
-               0 + offset, dst);
-        apply4(i0 | (1U << rev_wire_5), rev_wire_0, rev_wire_1, rev_wire_2,
-               rev_wire_3, rev_wire_4, 32 + offset, dst);
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply6(const std::size_t i0, const std::size_t rev_wire_0,
-                const std::size_t rev_wire_1, const std::size_t rev_wire_2,
-                const std::size_t rev_wire_3, const std::size_t rev_wire_4,
-                const std::size_t rev_wire_5, const std::size_t rev_wire_6,
-                const std::size_t offset, PrecisionT dst[]) const {
-        apply5(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3, rev_wire_4,
-               rev_wire_5, 0 + offset, dst);
-        apply5(i0 | (1U << rev_wire_6), rev_wire_0, rev_wire_1, rev_wire_2,
-               rev_wire_3, rev_wire_4, rev_wire_5, 64 + offset, dst);
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void apply7(const std::size_t i0, const std::size_t rev_wire_0,
-                const std::size_t rev_wire_1, const std::size_t rev_wire_2,
-                const std::size_t rev_wire_3, const std::size_t rev_wire_4,
-                const std::size_t rev_wire_5, const std::size_t rev_wire_6,
-                const std::size_t rev_wire_7, const std::size_t offset,
-                PrecisionT dst[]) const {
-        apply6(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3, rev_wire_4,
-               rev_wire_5, rev_wire_6, 0 + offset, dst);
-        apply6(i0 | (1U << rev_wire_7), rev_wire_0, rev_wire_1, rev_wire_2,
-               rev_wire_3, rev_wire_4, rev_wire_5, rev_wire_6, 128 + offset,
-               dst);
-    }
-
-    KOKKOS_INLINE_FUNCTION
-    void operator()(std::size_t k, PrecisionT dst[]) const {
-        if constexpr (num_wires == 0) {
-            std::size_t i0 = (k & parity[0]);
-            for (std::size_t i = 1; i < n_wires + 1; i++) {
-                i0 |= ((k << i) & parity[i]);
-            }
-            for (std::size_t inner_idx = 0; inner_idx < value_count;
-                 inner_idx++) {
-                std::size_t idx = i0;
-                for (std::size_t i = 0; i < n_wires; i++) {
-                    idx |= ((inner_idx & (one << i)) >> i) << rev_wires[i];
-                }
-                const PrecisionT rsv = real(arr(idx));
-                const PrecisionT isv = imag(arr(idx));
-                dst[inner_idx] += rsv * rsv + isv * isv;
-            }
-        }
-        if constexpr (num_wires == 1) {
-            const std::size_t i0 = ((k << 1U) & parity_1) | (k & parity_0);
-            apply0(i0, rev_wire_0, 0, dst);
-        }
-        if constexpr (num_wires == 2) {
-            const std::size_t i0 = ((k << 2U) & parity_2) |
-                                   ((k << 1U) & parity_1) | (k & parity_0);
-            apply1(i0, rev_wire_0, rev_wire_1, 0, dst);
-        }
-        if constexpr (num_wires == 3) {
-            const std::size_t i0 = ((k << 3U) & parity_3) |
-                                   ((k << 2U) & parity_2) |
-                                   ((k << 1U) & parity_1) | (k & parity_0);
-            apply2(i0, rev_wire_0, rev_wire_1, rev_wire_2, 0, dst);
-        }
-        if constexpr (num_wires == 4) {
-            const std::size_t i0 = ((k << 4U) & parity_4) |
-                                   ((k << 3U) & parity_3) |
-                                   ((k << 2U) & parity_2) |
-                                   ((k << 1U) & parity_1) | (k & parity_0);
-            apply3(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3, 0, dst);
-        }
-        if constexpr (num_wires == 5) {
-            const std::size_t i0 =
-                ((k << 4U) & parity_5) | ((k << 4U) & parity_4) |
-                ((k << 3U) & parity_3) | ((k << 2U) & parity_2) |
-                ((k << 1U) & parity_1) | (k & parity_0);
-            apply4(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3,
-                   rev_wire_4, 0, dst);
-        }
-        if constexpr (num_wires == 6) {
-            const std::size_t i0 =
-                ((k << 6U) & parity_6) | ((k << 5U) & parity_5) |
-                ((k << 4U) & parity_4) | ((k << 3U) & parity_3) |
-                ((k << 2U) & parity_2) | ((k << 1U) & parity_1) |
-                (k & parity_0);
-            apply5(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3,
-                   rev_wire_4, rev_wire_5, 0, dst);
-        }
-        if constexpr (num_wires == 7) {
-            const std::size_t i0 =
-                ((k << 7U) & parity_7) | ((k << 6U) & parity_6) |
-                ((k << 5U) & parity_5) | ((k << 4U) & parity_4) |
-                ((k << 3U) & parity_3) | ((k << 2U) & parity_2) |
-                ((k << 1U) & parity_1) | (k & parity_0);
-            apply6(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3,
-                   rev_wire_4, rev_wire_5, rev_wire_6, 0, dst);
-        }
-        if constexpr (num_wires == 8) {
-            const std::size_t i0 =
-                ((k << 8U) & parity_8) | ((k << 7U) & parity_7) |
-                ((k << 6U) & parity_6) | ((k << 5U) & parity_5) |
-                ((k << 4U) & parity_4) | ((k << 3U) & parity_3) |
-                ((k << 2U) & parity_2) | ((k << 1U) & parity_1) |
-                (k & parity_0);
-            apply7(i0, rev_wire_0, rev_wire_1, rev_wire_2, rev_wire_3,
-                   rev_wire_4, rev_wire_5, rev_wire_6, rev_wire_7, 0, dst);
         }
     }
 };
